% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{modularise_stan_files}
\alias{modularise_stan_files}
\title{Convert stan files to C++ modules in a package}
\usage{
modularise_stan_files(pkg = ".")
}
\arguments{
\item{pkg}{The directory containing the package.}
}
\description{
Converts stan files to C++ modules in a package.
}
